Layers are indexed from the django cache, that is populated by each layer check. Layers are also indexed in batch
