+++ hypermap/aggregator/utils.py
@@ -5,10 +5,13 @@ import re
 import sys
 import math
 import traceback
+import datetime
 from urlparse import urlparse
-
-from django.conf import settings
 from lxml import etree
+
+from django.utils.html import strip_tags
+from django.conf import settings
+
 from owslib.csw import CatalogueServiceWeb, CswRecord
 from owslib.wms import WebMapService
 from owslib.tms import TileMapService
@@ -372,3 +375,144 @@ def bbox2wktpolygon(bbox):
     maxy = float(bbox[3])
     return 'POLYGON((%.2f %.2f, %.2f %.2f, %.2f %.2f, %.2f %.2f, %.2f %.2f))' \
         % (minx, miny, minx, maxy, maxx, maxy, maxx, miny, minx, miny)
+
+
+def get_solr_date(pydate, is_negative):
+    """
+    Returns a date in a valid Solr format from a string.
+    """
+    # check if date is valid and then set it to solr format YYYY-MM-DDThh:mm:ssZ
+    try:
+        if isinstance(pydate, datetime.datetime):
+            solr_date = '%sZ' % pydate.isoformat()[0:19]
+            if is_negative:
+                print '***** This layer has a negative date'
+                solr_date = '-%s' % solr_date
+            return solr_date
+        else:
+            return None
+    except Exception:
+        return None
+
+
+def get_date(layer):
+    """
+    Returns a custom date representation. A date can be detected or from metadata.
+    It can be a range or a simple date in isoformat.
+    """
+    date = None
+    sign = '+'
+    date_type = 1
+    layer_dates = layer.get_layer_dates()
+    # we index the first date!
+    if layer_dates:
+        sign = layer_dates[0][0]
+        date = layer_dates[0][1]
+        date_type = layer_dates[0][2]
+    if date is None:
+        date = layer.created.date()
+    # layer date > 2300 is invalid for sure
+    # TODO put this logic in date miner
+    if date.year > 2300:
+        date = None
+    if date_type == 0:
+        date_type = "Detected"
+    if date_type == 1:
+        date_type = "From Metadata"
+    return get_solr_date(date, (sign == '-')), date_type
+
+
+def get_domain(url):
+    urlParts = urlparse(url)
+    hostname = urlParts.hostname
+    if hostname == "localhost":
+        return "Harvard"  # assumption
+    return hostname
+
+
+def layer2dict(layer):
+    """
+    Return a json representation for a layer.
+    """
+    logger = logging.getLogger("hypermap")
+    category = None
+    username = None
+    bbox = None
+    if not layer.has_valid_bbox():
+        message = 'There are not valid coordinates for layer id: %s' % layer.id
+        logger.error(message)
+    else:
+        bbox = [float(layer.bbox_x0), float(layer.bbox_y0), float(layer.bbox_x1), float(layer.bbox_y1)]
+        for proj in layer.service.srs.values():
+            if proj['code'] in ('102113', '102100'):
+                bbox = mercator_to_llbbox(bbox)
+        minX = bbox[0]
+        minY = bbox[1]
+        maxX = bbox[2]
+        maxY = bbox[3]
+        # coords hack needed by solr
+        if (minX < -180):
+            minX = -180
+        if (maxX > 180):
+            maxX = 180
+        if (minY < -90):
+            minY = -90
+        if (maxY > 90):
+            maxY = 90
+        wkt = "ENVELOPE({:f},{:f},{:f},{:f})".format(minX, maxX, maxY, minY)
+        halfWidth = (maxX - minX) / 2.0
+        halfHeight = (maxY - minY) / 2.0
+        area = (halfWidth * 2) * (halfHeight * 2)
+    domain = get_domain(layer.service.url)
+    if hasattr(layer, 'layerwm'):
+        category = layer.layerwm.category
+        username = layer.layerwm.username
+    abstract = layer.abstract
+    if abstract:
+        abstract = strip_tags(layer.abstract)
+    else:
+        abstract = ''
+    if layer.type == "WM":
+        originator = username
+    else:
+        originator = domain
+
+    layer_dict = {
+                    'id': layer.id,
+                    'type': 'Layer',
+                    'layer_id': layer.id,
+                    'name': layer.name,
+                    'title': layer.title,
+                    'layer_originator': originator,
+                    'service_id': layer.service.id,
+                    'service_type': layer.service.type,
+                    'layer_category': category,
+                    'layer_username': username,
+                    'url': layer.url,
+                    'reliability': layer.reliability,
+                    'recent_reliability': layer.recent_reliability,
+                    'last_status': layer.last_status,
+                    'is_public': layer.is_public,
+                    'availability': 'Online',
+                    'location': '{"layerInfoPage": "' + layer.get_absolute_url() + '"}',
+                    'abstract': abstract,
+                    'domain_name': layer.service.get_domain
+                    }
+
+    solr_date, date_type = get_date(layer)
+    if solr_date is not None:
+        layer_dict['layer_date'] = solr_date
+        layer_dict['layer_datetype'] = date_type
+    if bbox is not None:
+        layer_dict['min_x'] = minX
+        layer_dict['min_y'] = minY
+        layer_dict['max_x'] = maxX
+        layer_dict['max_y'] = maxY
+        layer_dict['area'] = area
+        layer_dict['bbox'] = wkt
+        srs_list = [srs.encode('utf-8') for srs in layer.service.srs.values_list('code', flat=True)]
+        layer_dict['srs'] = srs_list
+    if layer.get_tile_url():
+        layer_dict['tile_url'] = layer.get_tile_url()
+
+    return layer_dict
