+++ hypermap/aggregator/views.py
@@ -12,7 +12,7 @@ from django.contrib.auth.decorators import login_required
 
 from models import Service, Layer
 from tasks import (check_all_services, check_service, check_layer, remove_service_checks,
-                   index_service, index_all_layers, index_layer, clear_solr)
+                   index_service, index_all_layers, index_layer, index_cached_layers, clear_solr)
 from enums import SERVICE_TYPES
 
 from hypermap import celeryapp
@@ -201,6 +201,11 @@ def celery_monitor(request):
                 index_all_layers()
             else:
                 index_all_layers.delay()
+        if 'index_cached' in request.POST:
+            if settings.SKIP_CELERY_TASK:
+                index_cached_layers()
+            else:
+                index_cached_layers.delay()
         if 'clear_solr' in request.POST:
             if settings.SKIP_CELERY_TASK:
                 clear_solr()
